mcp_server/requirements.txt
bleak
screeninfo
aiohttp
sounddevice
numpy
scipy
python-dotenv 